Changed structure of company to store lists of values instead of lists of HistoricalDataResponse class Hopefully this will be easier to work with and faster

+++ Trading/HistoricalDataResponse.cs
@@ -9,21 +9,21 @@ namespace Trading
     public class HistoricalDataResponse
     {
         public string date { get; set; }
-        public double open { get; set; }
-        public double high { get; set; }
-        public double low { get; set; }
-        public double close { get; set; }
+        public float open { get; set; }
+        public float high { get; set; }
+        public float low { get; set; }
+        public float close { get; set; }
         public int volume { get; set; }
         public int unadjustedVolume { get; set; }
-        public double change { get; set; }
-        public double changePercent { get; set; }
-        public double vwap { get; set; }
+        public float change { get; set; }
+        public float changePercent { get; set; }
+        public float vwap { get; set; }
         public string label { get; set; }
-        public double changeOverTime { get; set; }
-        public double MovingAverageClose { get; set; }
+        public float changeOverTime { get; set; }
+        public float MovingAverageClose { get; set; }
         public int MovingAverageVolume { get; set; }
-        public double RelativePriceVolume { get; set; }
-        public double MovingRelativePriceVolume { get; set; }
+        public float RelativePriceVolume { get; set; }
+        public float MovingRelativePriceVolume { get; set; }
     }
